deleteOne now does not run if not connected to the database.

+++ server/src/middleware/database/db-connect.ts
@@ -156,6 +156,13 @@ export async function deleteOne<T extends Document>(
     schema: Schema,
     filter: Record<string, any>
 ) {
+    dbConnect();
+    if (mongoose.connection.readyState === 1) {
     const my_model = mongoose.model<T>(collection_name, schema);
     return await my_model.findOneAndDelete(filter) !== null;
+    }
+    else {
+        logger.error('[deleteOne] Could not delete. DB is disconnected');
+        return null;
+    }
 }
